refactor: move rating helper methods to utilities module
